// src/app/components/interfaces/orders.interfaces.ts
export interface infoInputs {
  plant:        string;
  no_parte:     string;
  descrip:      string;
  no_orden:     string;
  clave:        String;
  medida:       string;
  pieza_medida: number;
  articulo:     number;
  pieza_caja:   number;
  num_visual:   string;
  autorizo:     string;
  stock:number;

}
// ...
export interface selectorden {
  id_order:       number;
  folio:          string;
  nombre_cliente: string;
  ciudad:         string;
  sucursal:       string;
  ordenpro:       string;
  fecha_entrega:  string;
  fecha_creacion: Date;
  sem_entrega:    number;
  num_articulo:   number;
  num_parte:      string;
  descripcion:    string;
  num_orden_com:  string;
  clave:          string;
  medida:         string;
  pieza_medida:   number;
  hojas_cortar:   number;
  num_paquetes:   string;
  cantidad:       number;
  piezas_caja:    number;
  cajas_utilizar: number;
  elaboro:        string;
  autorizo:       string;
}
// ...
export interface avance {
  "COUNT(num_parte)": number;
  avance:number;
  prueba:number;
}

export interface avance2 {
  avance:number;
}


export interface stock {
  stock:number;

}